Improved RAM usage by changing double to float - Beware: Potential loss of precision

+++ Control.java
@@ -20,8 +20,8 @@ import java.io.IOException;
  */
 public class Control {
 
-	private static final double MOVE_LENGTH = 0.30;
-	private static final double ZOOM_LENGTH = 0.15;
+	private static final float MOVE_LENGTH = (float) 0.30;
+	private static final float ZOOM_LENGTH = (float) 0.15;
 	private static final String NAME = "Map"; //Name of the window containing the map.
 	private final File dataDir = new File(".", "data"); //Where control needs to look for the nodeFile and edgeFile
 	private final String nodeFile = "kdv_node_unload.txt"; //The nodes used to construct the graph
@@ -137,8 +137,8 @@ public class Control {
 				int newWidth = v.getCanvasWidth();
 				int newHeight = v.getCanvasHeight();
 				
-				double x_adjust = map.width*(((double)newWidth - oldWidth)/oldWidth);
-				double y_adjust = map.height*(((double)newHeight - oldHeight)/oldHeight);
+				float x_adjust = (float) (map.width*(((float)newWidth - oldWidth)/oldWidth));
+				float y_adjust = (float) (map.height*(((float)newHeight - oldHeight)/oldHeight));
 				
 				m.updateBounds(new Rectangle2D.Double(map.x, map.y - y_adjust, map.width + x_adjust,
 						map.height + y_adjust));
@@ -152,7 +152,7 @@ public class Control {
 		});
 	}
 
-	private Rectangle2D.Double zoomRect(double factor, boolean zoom, Rectangle2D.Double old){
+	private Rectangle2D.Double zoomRect(float factor, boolean zoom, Rectangle2D.Double old){
 		if(zoom){
 			return new Rectangle2D.Double(old.x + factor * old.width, //x is increased by the factor in proportion to the width
 					old.y + factor * old.height, //y is increased by the factor in proportion to the height
@@ -197,20 +197,20 @@ public class Control {
 		Rectangle2D.Double map = m.getBounds();
 		e.y = v.getCanvasHeight() - e.y;
 		// convert pixel to meters
-		double x_m = map.x + (e.getX() / (double) v.getCanvasWidth()) * map.width;
-		double y_m = map.y + (e.getY() / (double) v.getCanvasHeight()) * map.height;
+		float x_m = (float) (map.x + (e.getX() / (float) v.getCanvasWidth()) * map.width);
+		float y_m = (float) (map.y + (e.getY() / (float) v.getCanvasHeight()) * map.height);
 		return new Point2D.Double(x_m, y_m);
 	}
 
 	private void fixRatio(Rectangle2D.Double a, Rectangle2D.Double b){
-		double ratio = b.width / b.height;
+		float ratio = (float) (b.width / b.height);
 		if(a.width < a.height){
-			double temp = a.width;	
+			float temp = (float) a.width;	
 			a.width = ratio * a.height;
 			a.x = a.x - (a.width - temp) / 2;
 		}
 		else{
-			double temp = a.height;	
+			float temp = (float) a.height;	
 			a.height = a.width / ratio;
 			a.y = a.y - (a.height - temp) / 2;
 		}
